export types and interface, fix import
