+++ .eslintrc.json
@@ -1,6 +1,6 @@
 {
   "extends": ["next/core-web-vitals"],
-  "plugins": ["prettier"],
+  "plugins": ["prettier", "import"],
   "settings": {
     "import/resolver": {
       "typescript": {}
