+++ src/app/layout.tsx
@@ -20,7 +20,7 @@ export default function RootLayout({children}: {children: React.ReactNode}) {
       <head>
         <title>Create Next App</title>
         <meta name='description' content='Generated by create next app' />
-        <link rel='icon' href='/favicon.ico' />
+        <link rel='icon' href='./favicon.ico' />
       </head>
       <CustomThemeProvider>
         <Providers>
