+++ src/components/LimitSlider.tsx
@@ -1,8 +1,9 @@
 import {Slider, Typography, Box, useTheme} from '@mui/material';
+
 import React from 'react';
 
-import {setLimit, setPage, onHasMore} from '../redux/features/searchSlice';
-import {useAppDispatch, useAppSelector} from '../redux/hooks';
+import {setLimit, setPage, onHasMore} from '@/redux/features/searchSlice';
+import {useAppDispatch, useAppSelector} from '@/redux/hooks';
 
 const LimitSlider: React.FC = () => {
   const theme = useTheme();
